requirements.txt
# Core dependencies
flask==2.3.3
requests==2.31.0
python-dotenv==1.0.0

# Optional: For development and testing (uncomment if needed)
# pytest==7.4.2
# pytest-flask==1.2.0
# black==23.7.0
# flake8==6.0.0